Add Action to test 1-click on stacks PRs
# Problem
We need an automatic way to do Kubernetes rollout testing of non-doc Pull Requests made to the stacks directory. This should give feedback to contributors if their DigitalOcean Kubernetes 1-click will be deployed as expected before requesting a reviewer from a repo Code Owner. 
The testing should use the same mechanisms as production 1-clicks to give accurate feedback. Testing should block any subsequent reviews on the PR, and give verbose logging to properly inform reviewers and submitters. 
Code reviewers should be able to lean on this auto-testing to ensure the PR successfully does a rollout deployment. Additional testing should be later added to run integration tests to ensure the application is functioning as expected. 

# Proposed Solution